FridayDiscount does not apply any discount if it is "Thursday".

+++ src/main/java/org/example/Product.java
@@ -1,4 +1,3 @@
 package org.example;
 
-public record Product(String name, double price, int quantity) {
-}
+public record Product(String name, double price, double quantity) {}
